fix: add sfc component
